Replace `NameResolution` with `ImportBucket`
