+++ ohdlc/Cargo.toml
@@ -10,4 +10,5 @@ deref-derive = "0.1.0"
 lasso = { version = "0.7.2", features = ["multi-threaded"] }
 logos = "0.14.0"
 once_cell = "1.19.0"
+petgraph = "0.6.4"
 surotto = { git = "https://github.com/DasLixou/surotto" }
