+++ ohdlc/src/ir/mod.rs
@@ -1,6 +1,6 @@
+pub mod import_bucket;
 pub mod modules;
 pub mod name_lookup;
-pub mod name_resolution;
 pub mod registry;
 pub mod types;
